Configurar PropertySourcesPlaceholderConfigurer
