+++ cct-system/src/main/java/com/cct/config/ApplicationConfig.java
@@ -1,0 +1,17 @@
+package com.cct.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
+
+@Configuration
+@PropertySource("classpath:application.properties")
+public class ApplicationConfig {
+
+	@Bean
+    public static PropertySourcesPlaceholderConfigurer placeholderConfigurer() {
+        return new PropertySourcesPlaceholderConfigurer();
+    }
+
+}
