--- cct-system/src/main/java/com/cct/config/CctSystemApplication.java
@@ -1,14 +1,0 @@
-package com.cct.config;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
-
-@SpringBootApplication
-@ComponentScan("com.cct")
-public class CctSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CctSystemApplication.class, args);
-	}
-}
