+++ cct-system/src/main/java/com/cct/config/RedisConfig.java
@@ -3,7 +3,6 @@ package com.cct.config;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
@@ -11,31 +10,26 @@ import org.springframework.data.redis.core.RedisTemplate;
 @Configuration
 public class RedisConfig {
 
-//	@Value("${redis.hostname}")
-//    private String redisHostName;
+	@Value("${redis.hostname}")
+    private String redisHostName;
 
-//    @Value("${redis.port}")
-//    private int redisPort;
-	
-    @Bean
-    public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-        return new PropertySourcesPlaceholderConfigurer();
-    }
-    
+    @Value("${redis.port}")
+    private int redisPort;
+
 	@Bean
 	RedisConnectionFactory connectionFactory() {
 		JedisConnectionFactory jedisConnectionFactory = new JedisConnectionFactory();
 		jedisConnectionFactory.setUsePool(true);
-		jedisConnectionFactory.setHostName("localhost");
-		jedisConnectionFactory.setPort(6379);
+		jedisConnectionFactory.setHostName(redisHostName);
+		jedisConnectionFactory.setPort(redisPort);
 		return jedisConnectionFactory;
 	}
-	
+
 	@Bean
     RedisTemplate<Object, Object> redisTemplate() {
         RedisTemplate<Object, Object> redisTemplate = new RedisTemplate<Object, Object>();
         redisTemplate.setConnectionFactory(connectionFactory());
         return redisTemplate;
     }
-	
+
 }
