+++ cct-system/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/viagging
-spring.datasource.username=postgres
-spring.datasource.password=admin
+spring.datasource.username=pol_v4
+spring.datasource.password=pol123456$t8911
 
 redis.hostname=localhost
-redis.post=6379
+redis.port=6379
